refactor: structure in database (#77)
