+++ heliotrope/database/orm/table/__init__.py
@@ -1,0 +1,5 @@
+from heliotrope.database.orm.table.file import file_table
+from heliotrope.database.orm.table.gallleryinfo import galleryinfo_table
+from heliotrope.database.orm.table.tag import tag_table
+
+__all__ = ["file_table", "tag_table", "galleryinfo_table"]
